Change to stop crash on compile and run of example.
Change line 15 in example/main.go to take the address of the structure. This was being caught on line 11 of checker.go where the value is checked to ensure it is a pointer.
This make the example work again. Suspect that the rule validation was changed after the example was constructed.